First step toward going back to filesytem storage.
Trying to be a bit more organized now, I've recreated the filesystem-based
crawl, with subscriptions being stored in a subscriptions.opml file.  This
dispenses with the feed category symlinks, and gives a natural place to produce
a pre-cached HTML version of each feed entry for the reading interface.

The `skim.crawl` module should be functional, but `skim.subscriptions` and
`skim.entries` are pretty far out of alignment, with most functions raising
`NotImplementedError`.  Crawl times were very promising, and felt faster than
the Elasticsearch version.